Add useAppData hook
